main/postmarketos-base-ui-gnome: add gcr-ssh-agent to depends (MR 5120)
The SSH agent built into gnome-keyring has been disabled by default
in version 46, and replaced by the SSH agent implementation in gcr.
Install the new implementation to provide a graphical SSH agent
for all GNOME-based UIs.

Follow-up to aports!63893, which fixes this upstream and adds this
dependency to the upstream gnome package.

Fixes #2728.
[ci:skip-build]: already built successfully in CI