temp/gtk+3.0: remove (MR 5235)
Since we now install papers by default instead of Evince, every single
app we install from GNOME packages is either adaptive or GTK4.

The only real difference I have seen is that portals with Phosh are still
using GTK3. Given that Phosh allows to fit windows to the view, I would
consider this an acceptable regression, as there is a simple and common
workaround available.

Fixes #1479
[ci:skip-build]: already built successfully in CI