main/dtbtool-sprd: add dtc as a dependency as a fallback (MR 5275)
Currently, dtbtool-sprd has an argument -p which accepts the path to the
dtc binary. When it's not provided, it should use the dtc binary installed
in the system as a fallback. However, that's not the case. The following
error is reported:

Found file: test.dtb ... skip, failed to scan for 'sprd,sc-id = <' tag

...even though test.dtb has the `sprd,sc-id` property.

In order to mitigate such issues, install dtc as a dependency of the
package.