main/postmarketos-initramfs-minimal: don't hard-code libz path (MR 5610)
It's the work of the mkinitfs to fetch the library. This was initially
introduced in c7c2b2d02aeb255fb1dfc60e97603e4bdce74fd1 and its use-case
is non-existent today as osk-sdl is gone. In addition, the mkinitfs
rework Clayton did some time ago correctly detects libraries needed for
binaries, exactly so we don't need to pull this in manually!

[ci:skip-build]: already built successfully in CI