device-nokia-n900: lock.sh: query touchscreen instead of display (MR 5930)
lock.sh is only reliable when used with i3wm, where we disable gpio_keys from
being handled by X/i3wm. When this is not done (e.g. when testing a different
X11 UI), the script becomes unreliable since the display turns on due to the
SCREENLOCK key being toggled, and therefore only the display=on case is ever
run.

Improve reliability of the script by basing the decision making on the
enabled/disabled state of the touchscreen instead.