soc-qcom-sc8180x: remove nonexistent package from depends (MR 6019)
The qrtr-openrc package was removed in 94a777390c2e32e1acd8cb352fd8d1fbd8000299
but somehow still left in depends of this package.

Fix the errors like:

  ERROR: unable to select packages:
    qrtr-openrc (no such package):
      required by: soc-qcom-sc8180x-nonfree-firmware-openrc-1-r1[qrtr-openrc]

This is the one and only mention of qrtr-openrc package in pmaports.

P.S. It's very telling how many users sc8180x has, if this was unnoticed
for 8 months.

Signed-off-by: Alexey Min <alexey.min@gmail.com>