linux-postmarketos-qcom-sdm845: Improve work-around for call audio regressions (MR 6032)
Somehow enabling in-kernel PD mapper (introduced in 5b9f51b200dcb
"remoteproc: qcom: enable in-kernel PD mapper"), breaks the bottom mic
on xiaomi-beryllium (Poco F1) and distorts audio on OP6 (and possibly
OP6T). Temporarily disable it until the issue is resolved upstream.

Previously, the module's was modified to disable the in-kernel PD
mapper. This modification was a hack. Instead, modify the config so that
the in-kernel PD mapper is not built in the first place, which achieves
the same result.

Link: https://lore.kernel.org/linux-arm-msm/20241010074246.15725-1-johan+linaro@kernel.org/T
Link: pmaports#3353
Link: pmaports#3320
Link: pmaports#3245
Link: pmaports#3253

Signed-off-by: Frank Oltmanns <frank@oltmanns.dev>