linux-next: hotfix for missing driver in the build for qcom snps usb phy
The current linux-next packages are missing the driver for the eUSB PHY
which breaks USB on Qualcomm laptops. This happened because the driver
was moved out of the Qualcomm subdirectory and the config variable has
changed [1].. but the arm64 defconfig was not yet updated to match!

[1]: https://lore.kernel.org/all/20250504144527.1723980-4-ivo.ivanov.ivanov1@gmail.com/

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6538