device-xiaomi-{davinci,surya}: Move from testing to community (MR 6210)
The SM7150 devices have seen a lot of work put into them and are
generally very usable except for microphone/earpiece audio and cameras.
Both of these features are still work in progress and should be
available sometime in the future. They fulfil the requirements for
community in the Wiki, so let's promote them.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>