device-odroid-hc2: Force MBR/msdos partition table type
GPT paritions seem to interfere with the firmware baked into the SD card image;
both U-Boot and various Linux tools report the GPT partition table as broken
and fall back to the backup table. Notably, this makes the card difficult to
mount in another device. "Fixing" the partition table makes it refuse to boot.

Force the msdos partition type in deviceinfo to fix this.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6342