ci: avoid using alpine:edge on lint CI jobs
It is a lot more prone to breaking than stable, which is the
default. The reason to have them there is also not very clear.
The editor-config has been there since its introduction in
e7a81411137075919bcd4064fc63107faa9c7b51. It might be that at the time
it was not in community or in a stable release. The shellcheck-grep
has been there since 81872a59c1d7398c8f38d5a2102d189d472f6426, which
was already 6 years ago. Probably at the time our tooling was a lot
less integrated and stable in alpine as it is today, and everything
was running edge.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6471