ci: Always add build jobs, but disable by default
This sounds counter-productive and counter-intuitive, but hear me out:
this allows using `needs` to depend on those jobs, without further
templating downstream jobs.

In turn, those jobs depending on a disabled jobs, while they still exist
in the “graph” of the pipeline, are also transitively disabled.

This change, for all intents and purpose, is a no-op regarding the
amounts of jobs being sent to CI runners.

Also updates the comment for the placeholder job, since the wording
could lead to think that *having jobs* is sufficient. In reality, having
jobs that ran succesfully is necessary for the child pipeline to be
successful.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6399