Revert "ci/pytest: make sure tests run on correct branch (MR 6418)"
The wrong branch was not only used for pytest, but for all CI tests that
used pmbootstrap. I've fixed the bug in pmbootstrap now and added a
generic check to ensure we would catch a similar bug if we had one in
the future.

This reverts commit 7fd9e01f7f6b70c160bc58d7400cbd58f6fdf6ac.

Related: pmaports issue 3674
Related: pmbootstrap MR 2586
Related: ci-common MR 14

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6439