device-samsung-m3: use msdos partition type
If a custom bl1 is used on sdcard, then it should be embedded at
0x200 (512). If a GPT filesystem is used, then the GPT header uses
roughly the first 0x500 bytes, and hence bl1 overwrites part of the
GPT header in that configuration.

Change to msdos partition type so that the filesystem works fine
whether or not we use stock bl1 or custom one.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6442