device-shift-otter: remove initfs compression and kernel cmdline
The initfs compression was a leftover from when the device aport
was using the sdm845 mainline kernel (as we played with the otter
bringup there).

This resulted in causing the device to not boot.

Also remove the kernel cmdline as it is already handled by getty.

Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6455
[ci:skip-build]: already built successfully in CI