main/cros-keyboard-map: add systemd service and supporting files
Using systemd-tmpfiles to create the directory in /tmp seems to be the most systemd way to do it.
The config file for the systemd has a different format because the format of the openrc file doesn't lend itself to being used in a systemd service.

Thanks to @craftyguy for suggested fix.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6529