kconfigcheck: add netmount category to support NFS better
NFS allow the user to mount distant resources locally, and this is
specially great for mobiles devices, with limited disk space. The user
can couple this with FS-Cache, to limit the network impact. Kerberos
bring security to the NFS mounting.

This new category "netmout" make sure the kernels are built to support
this.