CI: test_aports_ui: fix systemd + armhf dep case
Fix the following error that now appears as the systemd-enabled
gnome-keyring package in pmaports will not be built for armhf:

  main/postmarketos-ui-xfce4/APKBUILD: package 'gnome-keyring' from _pmb_recommends not found for arch 'armhf'

This happened, because before this patch pmb.helpers.package.get() would
return the pmaports version of the package (!armhf) instead of the one
in the binary packages from alpine (alpine one is available for armhf),
and then later check the arches of the returned package and fail because
it is not enabled for armhf.

With the patch, the binary package from alpine repos is returned as the
pmaports one is ruled out due to not being available for armhf. The
latter arch check is not necessary anymore and gets removed.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6710