main/postmarketos-ui-os-installer: don't run os-installer and pmbootstrap as root
I'm not sure what the original motivation for doing this was, I would
guess just to get a minimal viable product working as soon as possible.
But based on my current testing, with these changes along with an
earlier commit that disabled the password requirement for privilege
escalation with doas, it is unnecessary and also solves at least one bug.

Closes https://gitlab.postmarketos.org/postmarketOS/pmaports/-/issues/3848

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/6787
[ci:skip-build]: already built successfully in CI